white_dune 1.956 update: minor bugfix: fixed X3D to C++ compiler

+++ src/MeshMorphingNode.cpp
@@ -74,7 +74,7 @@ MeshMorphingNode::write(int f, int indent, bool avoidUse)
 }
 
 void
-MeshMorphingNode::addToConvertedNodes(int flags)
+MeshMorphingNode::addToConvertedNodes(long flags)
 {
     if (m_already_converted)
         return;
